feat(Fortune cookie & item):
* LibFortuneCookie: fortune cookie mint & burn feature
* LibAppStorage:  modify fortune cookie mint time field - for an easier development experience.

+++ contracts/oracle_book_contracts/libs/LibAppStorage.sol
@@ -80,7 +80,7 @@ struct VRFRequestStatus {
 struct FortuneCookie {
     
     /** the generation time of this fortune cookie */
-    uint40 generateTime;
+    uint256 generateTime;
 
     /** encrypted data */
     bytes encryptedData;
